requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
python-dotenv
gunicorn
